UPSTREAM: coresight: etm4x: Add view comparator settings API to sysfs.
Currently it is not possible to view the current settings of a given
address comparator without knowing what type it is set to. For example, if
a comparator is set as an addr_start comparator, attempting to read
addr_stop for the same index will result in an error.

addr_cmp_view is added to allow the user to see the current settings of
the indexed address comparator without resorting to trial and error when
the set type is not known.

Signed-off-by: Mike Leach <mike.leach@linaro.org>
Reviewed-by: Leo Yan <leo.yan@linaro.org>
Signed-off-by: Mathieu Poirier <mathieu.poirier@linaro.org>
Link: https://lore.kernel.org/r/20191104181251.26732-11-mathieu.poirier@linaro.org
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit a578427de5b2bb7ac52f807505a6bea7fff93d4b)

BUG=b:149851136
TEST=See /sys/bus/coresight has devices on Trogdor

Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Change-Id: I944c90847f822199ce712e16dde9b34093b44957
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2163681
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>