UPSTREAM: ALSA: memalloc: Allow NULL device for SNDRV_DMA_TYPE_CONTINUOUS type
Currently we pass the artificial device pointer to the allocation
helper in the case of SNDRV_DMA_TYPE_CONTINUOUS for passing the GFP
flags.  But all common cases are the allocations with GFP_KERNEL, and
it's messy to put this in each place.

In this patch, the memalloc core helper is changed to accept the NULL
device pointer and it treats as the default mode, GFP_KERNEL, so that
all callers can omit the complex argument but just leave NULL.

Link: https://lore.kernel.org/r/20191105080138.1260-2-tiwai@suse.de
Signed-off-by: Takashi Iwai <tiwai@suse.de>
(cherry picked from commit 08422d2c559d1119b17c14e79b2d8ff89f46a356)
Signed-off-by: Chintan Patel <chintan.m.patel@intel.com>

BUG=b:147436144
TEST=Audio playback via Speaker, Headphone and HDMI/DP

Change-Id: I0b8f22c587606914d55857f16af5be4580fb51f5
Signed-off-by: Curtis Malainey <cujomalainey@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2029776
Reviewed-by: Sean Paul <seanpaul@chromium.org>