CHROMIUM: iwl7000: don't send PPAG command if disabled
we should not send the PPAG (Per-Platform Antenna Gain)
command to FW unless the platform has this ACPI table and it was
read and validated during the init flow. also no need to send the
command if the feature is disabled, so check if enabled before
sending, as if there is no valid table the feature is disabled.

Change-Id: I091fd10a29fcfabd47c97f2ff32bd8104d56db44
Signed-off-by: Gil Adam <gil.adam@intel.com>
iwl7000-tree: ff525fc7e325151c75697d225433148a0af5c24e