CHROMIUM: mtk-vcodec: Compute plane address with data_offset
With V4L2 MEMORY DMABUF API, it is necessary to set a file descriptor on
each plane. The way to pass an offset with a DmaBuf is not defined in V4L2
specification. This is the case as it is needed to specify offsets because
Chromium creates one DmaBuf for the planes.
This is HACK to abuse data_offset for the purpose.
This patch should be reverted after the appropriate way is defined.

BUG=b:158436469
BUG=chromium:901264
BUG=b:144135251
TEST=Kernel builds for Asurada.
Signed-off-by: Hirokazu Honda <hiroh@chromium.org>
Signed-off-by: Alexandre Courbot <acourbot@chromium.org>
Change-Id: I9ac3d0f1521bb02d700966a5aeab17caffee191d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2721869
Commit-Queue: Tzung-Bi Shih <tzungbi@chromium.org>
Reviewed-by: Tzung-Bi Shih <tzungbi@chromium.org>