UPSTREAM: ath10k: Handle "invalid" BDFs for msm8998 devices
When the BDF download QMI message has the end field set to 1, it signals
the end of the transfer, and triggers the firmware to do a CRC check.  The
BDFs for msm8998 devices fail this check, yet the firmware is happy to
still use the BDF.  It appears that this error is not caught by the
downstream drive by concidence, therefore there are production devices
in the field where this issue needs to be handled otherwise we cannot
support wifi on them.  So, attempt to detect this scenario as best we can
and treat it as non-fatal.

Signed-off-by: Jeffrey Hugo <jeffrey.l.hugo@gmail.com>
Reviewed-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Signed-off-by: Kalle Valo <kvalo@codeaurora.org>
(cherry picked from commit 319c2b71041f324e06dbee03c1011106cf746795)

BUG=b:152645955
TEST=Booted trogdor, observed wifi

Signed-off-by: Sujit Kautkar <sujitka@google.com>
Change-Id: Idadf3946854651938efbbe0ea0b78f87b392a779
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2171908
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Evan Green <evgreen@chromium.org>
Tested-by: Evan Green <evgreen@chromium.org>
Tested-by: Sujit Kautkar <sujitka@chromium.org>
Commit-Queue: Sujit Kautkar <sujitka@chromium.org>