CHROMIUM: iwl7000: pcie: detect the DMA bug and warn if it happens
Warn if the DMA bug is going to happen. We don't have a good
way of actually aborting in this case and we have workarounds
in place for the cases where it happens, but in order to not
be surprised add a safety-check and warn.

Change-Id: If5e419f6e7cff3a691b4434697e091b6db735652
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 73d573e6719b881ef95f0444c9ea6b27a0da8db2