BACKPORT: FROMLIST: platform: chrome: cros-ec: record event timestamp in the hard irq
To improve sensor timestamp precision, given EC and AP are in
different time domains, the AP needs to try to record the exact
moment an event was signalled to the AP by the EC as soon as
possible after it happens.

First thing in the hard irq is the best place for this.

Conflicts:
	drivers/platform/chrome/cros_ec_ishtp.c: spaces/tab

(am from https://patchwork.kernel.org/patch/11245749/)

BUG=b:67743747, chromium:944342
TEST=conpile

Acked-by: Jonathan Cameron <Jonathan.Cameron@kernel.org>
Acked-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>
Acked-by: Lee Jones <lee.jones@linaro.org>
Change-Id: Ided5e2457c8027a3e887072b1aebbd824ab2dd87
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>