BACKPORT: dma-mapping: treat dev->bus_dma_mask as a DMA limit
Using a mask to represent bus DMA constraints has a set of limitations.
The biggest one being it can only hold a power of two (minus one). The
DMA mapping code is already aware of this and treats dev->bus_dma_mask
as a limit. This quirk is already used by some architectures although
still rare.

With the introduction of the Raspberry Pi 4 we've found a new contender
for the use of bus DMA limits, as its PCIe bus can only address the
lower 3GB of memory (of a total of 4GB). This is impossible to represent
with a mask. To make things worse the device-tree code rounds non power
of two bus DMA limits to the next power of two, which is unacceptable in
this case.

In the light of this, rename dev->bus_dma_mask to dev->bus_dma_limit all
over the tree and treat it as such. Note that dev->bus_dma_limit should
contain the higher accessible DMA address.

Signed-off-by: Nicolas Saenz Julienne <nsaenzjulienne@suse.de>
Reviewed-by: Robin Murphy <robin.murphy@arm.com>
Signed-off-by: Christoph Hellwig <hch@lst.de>
(cherry picked from commit a7ba70f1787f977f970cd116076c6fce4b9e01cc)

Conflicts:
	arch/x86/pci/sta2x11-fixup.c
	kernel/dma/direct.c
(Yong: this patch only replace bus_dma_mask.
 1. don't change the original flow which don't have bus_dma_mask.
 2. Keep the current ARCH_ZONE_DMA_BITS
 3. The function "report_addr" is added to include/linux/dma-direct.h
    via cros kernel commit e0d55dc81fb636b1776a1f01dd6b2f98bc7534f2)

BUG=b:187155802
TEST=Build and boot

Signed-off-by: Yong Wu <yong.wu@mediatek.com>
Change-Id: I667833a2ce7af0edf54ee62be706bc294f6b7095
Signed-off-by: Flora Fu <flora.fu@mediatek.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2738864
Reviewed-by: Tzung-Bi Shih <tzungbi@chromium.org>
Commit-Queue: Tzung-Bi Shih <tzungbi@chromium.org>