CHROMIUM: iwl7000: mac80211: fix HE disablement
Recent patches disabled HE whenever VHT is disabled. This is wrong,
as it should be disabled if VHT is disable only on high band. This
is handled later in the assoc flow. This change broke connecting with
HE on low band.

Change-Id: Idcd6f9c3ed69d7f533765e2ab053ca738e58e698
Signed-off-by: Sara Sharon <sara.sharon@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 946c3fd8bf0025c34aa84021d4093d21554b36be