CHROMIUM: iwl7000: mac80211: he: set rx nss
In case of HE, the rx nss is taken from the HE capabilities.
If the supported nss capabilities that are reported by AP for the 11ax
mode in the HE Capabilities element are different from the nss
capabilities that are reported by AP for the vht mode in the vht
Capabilities element, need to use the lowest supported nss.

Change-Id: Iec80ef1407c6d01b0db8100fc182741ce35aeede
Signed-off-by: Tova Mussai <tova.mussai@intel.com>
iwl7000-tree: 31eccf53ed9b3af5f4397ed43af09801b685e855