CHROMIUM: iwl7000: use pointer arithmetic correctly when parsing wowlan status
v7 is a pointer to truct iwl_wowlan_status so when adding status_size to
the pointer, we are essentially doing v7 + sizeof(v7) * status_size,
while status_size is the size in bytes.  So cast the pointer to a u8
pointer in order to make the addtion correctly.

Change-Id: I6f343df48a75a97159e237a6a29ab0652afc85d0
Signed-off-by: Mordechay Goodstein <mordechay.goodstein@intel.com>
iwl7000-tree: ef6c729549d0739efb04cd27345135988db2fc12