BACKPORT: arm64: Add initial support for E0PD
Kernel Page Table Isolation (KPTI) is used to mitigate some speculation
based security issues by ensuring that the kernel is not mapped when
userspace is running but this approach is expensive and is incompatible
with SPE.  E0PD, introduced in the ARMv8.5 extensions, provides an
alternative to this which ensures that accesses from userspace to the
kernel's half of the memory map to always fault with constant time,
preventing timing attacks without requiring constant unmapping and
remapping or preventing legitimate accesses.

Currently this feature will only be enabled if all CPUs in the system
support E0PD, if some CPUs do not support the feature at boot time then
the feature will not be enabled and in the unlikely event that a late
CPU is the first CPU to lack the feature then we will reject that CPU.

This initial patch does not yet integrate with KPTI, this will be dealt
with in followup patches.  Ideally we could ensure that by default we
don't use KPTI on CPUs where E0PD is present.

Signed-off-by: Mark Brown <broonie@kernel.org>
Reviewed-by: Suzuki K Poulose <suzuki.poulose@arm.com>
[will: Fixed typo in Kconfig text]
Signed-off-by: Will Deacon <will@kernel.org>

(cherry picked from commit 3e6c69a058deaa50d33c3dac36cde80b4ce590e8)

Conflicts:
   arch/arm64/include/asm/cpucaps.h
...resolve the same way upstream did in merge commit ab3906c53144
("Merge branch 'for-next/errata' into for-next/core")

BUG=b:144577102
TEST=Future patches pick cleanly

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: I9efaaf3f8274af7c143df8f01fe708c000e998c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2204710
Reviewed-by: Stephen Boyd <swboyd@chromium.org>