UPSTREAM: cpuidle: Fix cpuidle_driver_state_disabled()
It turns out that cpuidle_driver_state_disabled() can be called
before registering the cpufreq driver on some platforms, which
was not expected when it was introduced and which leads to a NULL
pointer dereference when trying to walk the CPUs associated with
the given cpuidle driver.

Fix the problem by making cpuidle_driver_state_disabled() check if
the driver's mask of CPUs associated with it is present and to set
CPUIDLE_FLAG_UNUSABLE for the given idle state in the driver's states
list if that is not the case to cause __cpuidle_register_device() to
set CPUIDLE_STATE_DISABLED_BY_DRIVER for that state for all cpuidle
devices registered by it later.

Fixes: cbda56d5fefc ("cpuidle: Introduce cpuidle_driver_state_disabled() for driver quirks")
Reported-by: Daniel Lezcano <daniel.lezcano@linaro.org>
Tested-by: Daniel Lezcano <daniel.lezcano@linaro.org>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
(cherry picked from commit b0142d66f4edb8578b7772d6d7ad731836b82ddb)

BUG=b:146144745
TEST=Kernel boots on volteer

Signed-off-by: Ed Baker <edward.baker@intel.com>
Change-Id: I20e64ec44dfc21e8bd92d95f9d242b7099c2a900
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2092317
Tested-by: Minh QuangX Nguyen <minh.quangx.nguyen@intel.com>
Reviewed-by: Alex Levin <levinale@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>