CHROMIUM: iwl7000: pcie: rename L0S_ENABLED bit to L0S_DISABLED
This bit has been misnamed since the initial implementation of the
driver.  The correct semantics is that setting this bit disables L0S
states, and we already clearly use it as such in the code.  Rename it
to avoid confusion.

Change-Id: Ie2097de87f0a5c71d2d05bcf6348a21b04d7ec1f
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 65c1d91eeb9cab3fe7588177b5c5f1bf70f67cc5