CHROMIUM: iwl7000: mvm: fix SKB leak on invalid queue
It used to be the case that if we got here, we wouldn't warn
but instead allocate the queue (DQA). With using the mac80211
TXQs model this changed, and we really have nothing to do with
the frame here anymore, hence the warning now.

However, clearly we missed in coding & review that this is now
a pure error path and leaks the SKB if we return 0 instead of
an indication that the SKB needs to be freed. Fix this.

Change-Id: Iaeadd3e950a5c58a9b40dd4c40416544370fd8ee
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 2ca3dbd6f60fe9480f053676d8bc6f58af418fae