UPSTREAM: ACPI: processor: Export function to claim _CST control
The intel_idle driver will be modified to use ACPI _CST subsequently
and it will need to notify the platform firmware of that if
acpi_gbl_FADT.cst_control is set, so add a routine for this purpose,
acpi_processor_claim_cst_control(), to acpi_processor.c (so that it
is always present which is required by intel_idle) and export it
to allow the ACPI processor driver (which is modular) to call it.

No intentional functional impact.

Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
(cherry picked from commit bc94638886ab21f8247d3f7f39573d3feb7d8284)

BUG=b:146144745
TEST=Kernel boots on volteer

Signed-off-by: Ed Baker <edward.baker@intel.com>
Change-Id: I6a5543e6bfd1515b5b47c1be6f1f513c4735cff1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2092319
Tested-by: Minh QuangX Nguyen <minh.quangx.nguyen@intel.com>
Reviewed-by: Alex Levin <levinale@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>