UPSTREAM: intel_idle: Use ACPI _CST on server systems
In many cases, especially on server systems, it is desirable to avoid
enabling C-states that have been disabled in the platform firmware
(BIOS) setup, except for C1E.

As a rule, the C-states disabled this way are not listed by ACPI
_CST, so if that is used by intel_idle along with the specific
table of C-states that it has for the given processor, the C-states
disabled through the platform firmware will not be enabled by default
by intel_idle.

Accordingly, set the use_acpi flag (introduced previously) in all
server processor profiles defined in intel_idle (so as to make it use
ACPI _CST to decide which C-states to enable by default) and set
the CPUIDLE_FLAG_ALWAYS_ENABLE flag (also introduced previously)
for C1E in all C-states tables in intel_idle that contain C1 too
(so that C1E is enabled regardless of whether or not it is listed
by ACPI _CST).

Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
(cherry picked from commit e6d4f08a677654385869ba0c39d7c9ceec47e5c5)

BUG=b:146144745
TEST=Kernel boots on volteer

Signed-off-by: Ed Baker <edward.baker@intel.com>
Change-Id: I7f2dfb53224f811dde58db8acd83e02e6e078889
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2092331
Tested-by: Minh QuangX Nguyen <minh.quangx.nguyen@intel.com>
Reviewed-by: Alex Levin <levinale@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>