UPSTREAM: media: v4l2: abstract timeval handling in v4l2_buffer
As a preparation for adding 64-bit time_t support in the uapi,
change the drivers to no longer care about the format of the
timestamp field in struct v4l2_buffer.

The v4l2_timeval_to_ns() function is no longer needed in the
kernel after this, but there is userspace code relying on
it to be part of the uapi header.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
[hverkuil-cisco@xs4all.nl: replace spaces by tabs]
Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>
(cherry picked from commit 77cdffcb0bfb87fe3645894335cb8cb94917e6ac)

BUG=b:196080792, b:196080793
TEST=v4l2-compliance

Signed-off-by: Atanas Filipov <afilipov@mm-sol.com>
Change-Id: Iedde460a4de0a177ec48dc42e8e97791c99bcef4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3573927
Reviewed-by: Ricardo Ribalda <ribalda@chromium.org>
Commit-Queue: Ricardo Ribalda <ribalda@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Tested-by: Ricardo Ribalda <ribalda@chromium.org>