CHROMIUM: iwl7000: remove lar_disable module parameter
This is an old parameter that was used supposed to be used only when
LAR was still under development.  It should not be used anymore, but,
since it's available, end-users have been mangling with it
unnecessarily.  In some cases it can cause problems because when LAR
is supported the driver and the firmware do not expect it to be
disabled.

Change-Id: Ic96242ea501b17908a63f849fccb075cbda695a4
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 0371decf3567f02645d1d80cf0128bc162d55601