UPSTREAM: mac80211: Fix setting txpower to zero
With multiple VIFS ath10k, and probably others, tries to find the
minimum txpower for all vifs and uses that when setting txpower in
the firmware.

If a second vif is added and starts to scan, it's txpower is not
initialized yet and it set to zero.

ath10k had a patch to ignore zero values, but then it is impossible
to actually set txpower to zero.

So, instead initialize the txpower to INT_MIN in mac80211, and let
drivers know that means the power has not been set and so should
be ignored.

This should fix regression in:

commit 88407beb1b1462f706a1950a355fd086e1c450b6
Author: Ryan Hsu <ryanhsu@qca.qualcomm.com>
Date:   Tue Dec 13 14:55:19 2016 -0800

    ath10k: fix incorrect txpower set by P2P_DEVICE interface

Tested on ath10k 9984 with ath10k-ct firmware.

Signed-off-by: Ben Greear <greearb@candelatech.com>
Link: https://lore.kernel.org/r/20191217183057.24586-1-greearb@candelatech.com
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
(cherry picked from commit db6d9e9e8b48b7ab68c61553eb5fa68534dd0fde)

BUG=b:152645955
TEST=Booted trogdor, observed wifi

Signed-off-by: Sujit Kautkar <sujitka@chromium.org>
Change-Id: Ic987c28064f7fb61a644840b1e07c7610474bff6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2208064
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Abhishek Kumar <kuabhs@google.com>
Tested-by: Evan Green <evgreen@chromium.org>