UPSTREAM: ALSA: control: Fix incompatible protocol error
The recent change to bump the ALSA control API protocol version from
2.0.7 to 2.1.0 caused a regression on user-space; while the user-space
expects both the major and the minor versions to be identical with the
supported numbers, we changed the minor number from 0 to 1.

For recovering from the incompatibility, this patch changes the
protocol version again to 2.0.8, which is compatible, but yet higher
than the original number 2.0.7, indicating that the protocol change.

Fixes: bd3eb4e87eb3 ("ALSA: ctl: bump protocol version up to v2.1.0")
Reported-by: Paul Menzel <pmenzel@molgen.mpg.de>
Tested-by: Paul Menzel <pmenzel@molgen.mpg.de>
Link: https://lore.kernel.org/r/s5h1rsr769i.wl-tiwai@suse.de
Signed-off-by: Takashi Iwai <tiwai@suse.de>
(cherry picked from commit a103a3989993859091a26936b9337a8e09330fc3)
Signed-off-by: Chintan Patel <chintan.m.patel@intel.com>

BUG=b:155633508
TEST=No audio regression on TGL/JSL devices

Change-Id: I066a6f69bb12a505ca76ebdf481e8fdcc500c948
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2427220
Tested-by: Chintan M Patel <chintan.m.patel@intel.com>
Commit-Queue: Curtis Malainey <cujomalainey@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>