UPSTREAM: dt-bindings: nvmem: new optional property wp-gpios
Several memories have a write-protect pin, that when pulled high, it
blocks the write operation.

On some boards, this pin is connected to a GPIO and pulled high by
default, which forces the user to manually change its state before
writing.

Instead of modifying all the memory drivers to check this pin, make
the NVMEM subsystem check if the write-protect GPIO being passed
through the nvmem_config or defined in the device tree and pull it
low whenever writing to the memory.

Add a new optional property to the device tree binding document, which
allows to specify the GPIO line to which the write-protect pin is
connected.

Signed-off-by: Khouloud Touil <ktouil@baylibre.com>
Reviewed-by: Linus Walleij <linus.walleij@linaro.org>
Reviewed-by: Rob Herring <robh@kernel.org>
Signed-off-by: Bartosz Golaszewski <bgolaszewski@baylibre.com>
(cherry picked from commit 14f4957313fbf575c7bbd37d45cb148e11fdbc20)

BUG=b:110491551
TEST=Use efuse driver

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: I20c9beb3bfe37ffd5273e61edd32cf7dce9aa1c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2222570
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Evan Green <evgreen@chromium.org>