BACKPORT: io_uring: remove punt of short reads to async context
We currently punt any short read on a regular file to async context,
but this fails if the short read is due to running into EOF. This is
especially problematic since we only do the single prep for commands
now, as we don't reset kiocb->ki_pos. This can result in a 4k read on
a 1k file returning zero, as we detect the short read and then retry
from async context. At the time of retry, the position is now 1k, and
we end up reading nothing, and hence return 0.

Instead of trying to patch around the fact that short reads can be
legitimate and won't succeed in case of retry, remove the logic to punt
a short read to async context. Simply return it.

Signed-off-by: Jens Axboe <axboe@kernel.dk>
(cherry picked from commit eacc6dfaea963ef61540abb31ad7829be5eff284)

Conflicts:
- fs/io_uring.c (minor context diff)

BUG=chromium:901139
TEST=build

Signed-off-by: Keiichi Watanabe <keiichiw@chromium.org>
Change-Id: I3137a4842981cadb6e55366fb5e38e7d90b87287
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2655822
Reviewed-by: Dylan Reid <dgreid@chromium.org>