UPSTREAM: iommu/io-pgtable-arm: Ensure ARM_64_LPAE_S2_TCR_RES1 is unsigned
ARM_64_LPAE_S2_TCR_RES1 is intended to map to bit 31 of the VTCR register,
which is required to be set to 1 by the architecture. Unfortunately, we
accidentally treat this as a signed quantity which means we also set the
upper 32 bits of the VTCR to one, and they are required to be zero.

Treat ARM_64_LPAE_S2_TCR_RES1 as unsigned to avoid the unwanted
sign-extension up to 64 bits.

Cc: Robin Murphy <robin.murphy@arm.com>
Signed-off-by: Will Deacon <will@kernel.org>
(cherry picked from commit 6f932ad369a3c3f853ffc5d93de9a73420e862b1)

BUG=b:150320512
TEST=Boot trogdor with wifi/modem

Signed-off-by: Evan Green <evgreen@chromium.org>
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: I5a9b3d80119a260bb3a5ac324b4c92931e3f713c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2169196
Reviewed-by: Sean Paul <seanpaul@chromium.org>