UPSTREAM: iommu/io-pgtable-arm: Rationalise VTCR handling
Commit 05a648cd2dd7 ("iommu/io-pgtable-arm: Rationalise TCR handling")
reworked the way in which the TCR register value is returned from the
io-pgtable code when targetting the Arm long-descriptor format, in
preparation for allowing page-tables to target TTBR1.

As it turns out, the new interface is a lot nicer to use, so do the same
conversion for the VTCR register even though there is only a single base
register for stage-2 translation.

Cc: Robin Murphy <robin.murphy@arm.com>
Signed-off-by: Will Deacon <will@kernel.org>
(cherry picked from commit ac4b80e5b9d0ecf906300d79e4dc4df5526579a8)

BUG=b:150320512
TEST=Boot trogdor with wifi/modem

Signed-off-by: Evan Green <evgreen@chromium.org>
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: I59ca0eda1dd55758cd3bcd4a4abf440aaa991e41
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2169199
Reviewed-by: Sean Paul <seanpaul@chromium.org>