CHROMIUM: config: arm/arm64: Turn on uclamp
We live in "The Future (TM)" and uclamp is the way to tweak
power/performance in the future.  As per <https://crbug.com/1041117>
it replaces what was done in the past, AKA using schedutil.

This is:
  echo 'CONFIG_UCLAMP_TASK=y' >> chromeos/config/arm64/common.config
  echo 'CONFIG_UCLAMP_TASK=y' >> chromeos/config/armel/common.config
  echo 'CONFIG_UCLAMP_TASK_GROUP=y' >> chromeos/config/arm64/common.config
  echo 'CONFIG_UCLAMP_TASK_GROUP=y' >> chromeos/config/armel/common.config
  ./chromeos/scripts/kernelconfig olddefconfig
...atop an already-normalized config.

Note that I believe we don't use schedutil on x86 machines, so
enabling this only for arm/arm64 configs.

BUG=chromium:1041117
TEST=`find /sys/ | grep -i clamp` shows something now

Change-Id: I6948f5af5936a89f49c600fb3c966b2990e85f4a
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1997053
Commit-Queue: Guenter Roeck <groeck@chromium.org>
Commit-Queue: Sonny Rao <sonnyrao@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>